+++ docs/src/index.js
@@ -2,7 +2,7 @@
 import App from './components/app';
 
 // Material Components
-import '../../Typography/style.css';
+import '../../ts/Typography/style.css';
 import '../../Theme/style.css';
 
 // Styles
